fix new clippy warnings

+++ src/util.rs
@@ -26,7 +26,7 @@ fn get_hash<T: ?Sized + Hash, H: Hasher + Default>(data: &T) -> (u32, u32) {
 pub fn get_alt_index<H: Hasher + Default>(fp: Fingerprint, i: usize) -> usize {
     let (_, index_hash) = get_hash::<_, H>(&fp.data);
     let alt_i = index_hash as usize;
-    (i ^ alt_i) as usize
+    i ^ alt_i
 }
 
 impl FaI {
